+++ djangae/contrib/backup/tasks.py
@@ -35,8 +35,8 @@ def backup_datastore(bucket=None, kinds=None):
         return
 
     # make sure no blacklisted entity kinds are included in our export
-    valid_models = _get_valid_export_models(kinds)
-    if not valid_models:
+    valid_kinds = _get_valid_export_kinds(kinds)
+    if not valid_kinds:
         logger.warning("No whitelisted entity kinds to export.")
         return
 
@@ -45,7 +45,7 @@ def backup_datastore(bucket=None, kinds=None):
     body = {
         'outputUrlPrefix': get_backup_path(bucket),
         'entityFilter': {
-            'kinds': valid_models,
+            'kinds': valid_kinds,
         }
     }
     app_id = app_identity.get_application_id()
@@ -53,16 +53,17 @@ def backup_datastore(bucket=None, kinds=None):
     request.execute()
 
 
-def _get_valid_export_models(kinds=None):
+def _get_valid_export_kinds(kinds=None):
     """Make sure no blacklist models are included in our backup export."""
     excluded_models = get_backup_setting("EXCLUDE_MODELS", required=False, default=[])
     excluded_apps = get_backup_setting("EXCLUDE_APPS", required=False, default=[])
 
-    models_to_backup = []
+    to_backup = []
     for model in apps.get_models(include_auto_created=True):
         app_label = model._meta.app_label
         object_name = model._meta.object_name
         model_def = "{}_{}".format(app_label, object_name.lower())
+        db_table = model._meta.db_table
 
         if app_label in excluded_apps:
             logger.info(
@@ -71,7 +72,9 @@ def _get_valid_export_models(kinds=None):
             )
             continue
 
-        if model_def in excluded_models:
+        # Exclude the models if either the model label or datastore kind
+        # is listed.
+        if model_def in excluded_models or db_table in excluded_models:
             logger.info(
                 "Not backing up %s as it is blacklisted in DJANGAE_BACKUP_EXCLUDE_MODELS",
                 model_def
@@ -79,14 +82,17 @@ def _get_valid_export_models(kinds=None):
             continue
 
         logger.info("%s added to list of models to backup", model_def)
-        models_to_backup.append(model_def)
+        to_backup.append((model_def, db_table))
 
-    # if kinds we explcitly provided by the caller, we only return those
+    # If kinds we explcitly provided by the caller, we only return those
     # already validated by our previous checks
     if kinds:
-        models_to_backup = [model for model in models_to_backup if model in kinds]
-
-    return models_to_backup
+        return [
+            kind for (_model_def, kind) in to_backup
+            if _model_def in kinds or kind in kinds
+        ]
+    else:
+        return [kind for (_model_def, kind) in to_backup]
 
 
 def _get_service():